preparing --client-only dev/build process, added babel-plugin-macros
