+++ babel.config.js
@@ -18,6 +18,7 @@ module.exports = {
         '@babel/plugin-proposal-class-properties',
         '@babel/plugin-proposal-optional-chaining',
         '@babel/plugin-syntax-dynamic-import',
+        'macros',
     ],
     env: {
         test: {
