+++ config/webpack.config.js/template.html
@@ -1,11 +1,10 @@
 <!DOCTYPE html>
-<html lang="en">
+<html>
     <head>
         <meta charset="utf-8" />
         <!-- <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" /> -->
         <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
         <meta name="theme-color" content="#000000" />
-        <!-- <link rel="manifest" href="%PUBLIC_URL%/manifest.json" /> -->
         <title></title>
     </head>
     <body>
