+++ package.json
@@ -54,7 +54,7 @@
     "@babel/cli": "^7.2.3",
     "@babel/core": "^7.2.2",
     "@babel/plugin-proposal-class-properties": "^7.3.0",
-    "@babel/plugin-proposal-object-rest-spread": "^7.3.1",
+    "@babel/plugin-proposal-object-rest-spread": "^7.3.2",
     "@babel/plugin-proposal-optional-chaining": "^7.2.0",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
     "@babel/plugin-transform-modules-commonjs": "^7.2.0",
@@ -66,15 +66,16 @@
     "@svgr/webpack": "^4.1.0",
     "autoprefixer": "^9.4.7",
     "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.0.0",
+    "babel-jest": "^24.1.0",
     "babel-loader": "^8.0.5",
+    "babel-plugin-macros": "^2.4.5",
     "babel-plugin-named-asset-import": "^0.3.0",
     "case-sensitive-paths-webpack-plugin": "^2.2.0",
     "cross-env": "^5.2.0",
     "css-hot-loader": "^1.4.3",
     "css-loader": "^2.1.0",
     "enzyme": "^3.8.0",
-    "enzyme-adapter-react-16": "^1.8.0",
+    "enzyme-adapter-react-16": "^1.9.1",
     "eslint": "^5.13.0",
     "eslint-config-prettier": "^4.0.0",
     "eslint-config-wiremore": "^2.3.0",
@@ -83,16 +84,16 @@
     "eslint-plugin-import": "^2.16.0",
     "eslint-plugin-prettier": "^3.0.1",
     "eslint-plugin-react": "^7.12.4",
-    "eslint-plugin-react-hooks": "^0.0.0",
+    "eslint-plugin-react-hooks": "^1.0.1",
     "eslint-plugin-security": "^1.4.0",
     "file-loader": "^3.0.1",
-    "flow-bin": "^0.92.0",
+    "flow-bin": "^0.92.1",
     "html-webpack-plugin": "^3.2.0",
     "husky": "^1.3.1",
     "i18next-parser": "^0.13.0",
     "install-deps-postmerge": "^1.0.2",
-    "jest": "^24.0.0",
-    "lint-staged": "^8.1.1",
+    "jest": "^24.1.0",
+    "lint-staged": "^8.1.3",
     "mini-css-extract-plugin": "^0.5.0",
     "nodemon": "^1.18.9",
     "opn-cli": "^4.0.0",
@@ -104,8 +105,8 @@
     "postcss-loader": "^3.0.0",
     "postcss-nested": "^4.1.1",
     "postcss-normalize": "^7.0.1",
-    "prettier": "^1.16.3",
-    "puppeteer": "^1.12.1",
+    "prettier": "^1.16.4",
+    "puppeteer": "^1.12.2",
     "react-dev-utils": "^7.0.1",
     "redux-mock-store": "^1.5.3",
     "regenerator-runtime": "^0.13.1",
@@ -116,9 +117,9 @@
     "stylelint-order": "^2.0.0",
     "stylelint-prettier": "^1.0.6",
     "url-loader": "^1.1.2",
-    "webpack": "^4.29.0",
+    "webpack": "^4.29.3",
     "webpack-bundle-analyzer": "^3.0.3",
-    "webpack-cli": "^3.2.1",
+    "webpack-cli": "^3.2.3",
     "webpack-dev-server": "^3.1.14",
     "webpack-hot-middleware": "^2.24.3",
     "webpack-manifest-plugin": "^2.0.4",
